+++ testproject/Assets/Tests/Runtime/Support/NetworkVariableInitOnNetworkSpawn.cs
@@ -8,7 +8,8 @@ namespace TestProject.RuntimeTests.Support
         public NetworkVariable<int> Variable = new NetworkVariable<int>();
         public static bool NetworkSpawnCalledOnServer;
         public static bool NetworkSpawnCalledOnClient;
-        public static bool OnValueChangedCalledOnClient = false;
+        public static bool OnValueChangedCalledOnClient;
+        public static int ExpectedSpawnValueOnClient;
 
         private void Awake()
         {
@@ -25,7 +26,6 @@ namespace TestProject.RuntimeTests.Support
 
         public override void OnNetworkSpawn()
         {
-            Assert.IsFalse(OnValueChangedCalledOnClient);
             base.OnNetworkSpawn();
             if (IsServer)
             {
@@ -35,8 +35,8 @@ namespace TestProject.RuntimeTests.Support
             else
             {
                 NetworkSpawnCalledOnClient = true;
+                Assert.AreEqual(ExpectedSpawnValueOnClient, Variable.Value);
             }
-            Assert.AreEqual(5, Variable.Value);
         }
     }
 }